% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_short_string.R
\name{as_short_string}
\alias{as_short_string}
\title{Converts any R object to a descriptive string}
\usage{
as_short_string(x, num_format = "\%.4g", trunc_width = 15L)
}
\arguments{
\item{x}{[any]\cr
The object.}

\item{num_format}{[`character(1)]\cr
Used to format numerical scalars via \code{\link[base:sprintf]{base::sprintf()}}.
Default is "\%.4g".}

\item{trunc_width}{[\code{integer(1)}]\cr
Truncate strings to width \code{trunc_width}.
Default is 15.}
}
\value{
[`character(1)].
}
\description{
This function is intended to be convert any R object to a short descriptive string,
e.g. in \code{\link[base:print]{base::print()}} functions.

The following ruleset applies:
\itemize{
\item \code{x} is \code{atomic} with length 0 or 1: printed as-is.
\item \code{x} is of length greater than 1, they are first collapsed with ",", and the resulting string.
is then truncated to \code{trunc_width}.
\item \code{x} is an expression: converted to character.
\item Otherwise: print their class.
}

If \code{x} is a list, the above rules are applied (non-recursively) to its elements.
}
\examples{
as_short_string(list(a = 1, b = NULL, "foo", c = 1:10))
}
