% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callback.R
\name{assert_callback}
\alias{assert_callback}
\alias{assert_callbacks}
\title{Assertions for Callbacks}
\usage{
assert_callback(callback, null_ok = FALSE)

assert_callbacks(callbacks)
}
\arguments{
\item{callback}{(\link{Callback}).}

\item{null_ok}{(\code{logical(1)})\cr
If \code{TRUE}, \code{NULL} is allowed.}

\item{callbacks}{(list of \link{Callback}).}
}
\value{
\link{Callback} | List of \link{Callback}s.
}
\description{
Assertions for \link{Callback} class.
}
