% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_ro_binding.R
\name{assert_ro_binding}
\alias{assert_ro_binding}
\title{Assertion for Active Bindings in R6 Classes}
\usage{
assert_ro_binding(rhs)
}
\arguments{
\item{rhs}{(\code{any})\cr
If not missing, an exception is raised.}
}
\value{
Nothing.
}
\description{
This assertion is intended to be called in active bindings of an
\link[R6:R6Class]{R6::R6Class} which does not allow assignment.
If \code{rhs} is not missing, an exception is raised.
}
