% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the First Letter of Strings}
\usage{
capitalize(str)
}
\arguments{
\item{str}{(\code{character()}).}
}
\value{
Character vector, same length as \code{str}.
}
\description{
Takes a character vector and changes the first letter of each element to
uppercase.
}
\examples{
capitalize("foo bar")
}
