% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_cli.R
\name{cat_cli}
\alias{cat_cli}
\title{Function to transform message to output}
\usage{
cat_cli(expr)
}
\arguments{
\item{expr}{(\code{any})\cr
Expression that calls \verb{cli_*} methods, \code{\link[base:cat]{base::cat()}} or\code{\link[base:print]{base::print()}} to format an object's printout.}
}
\description{
Wrapper around \code{\link[cli:cli_format_method]{cli::cli_format_method()}}.
Uses \code{\link[base:cat]{base::cat()}} to transform the printout from a message to an output with a line break.
}
\examples{
cat_cli({
  cli::cli_h1("Heading")
  cli::cli_li(c("x", "y"))
})
}
