% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Chunk elements of vectors into blocks of nearly equal size.}
\usage{
chunk(x, n_chunks = NULL, chunk_size = NULL, shuffle = TRUE)
}
\arguments{
\item{x}{(\code{vector()}):
Vector to be split into chunks.}

\item{n_chunks}{(\code{integer(1)}):
Requested number of chunks.
Mutually exclusive with \code{chunk_size} and \code{props}.}

\item{chunk_size}{(\code{integer(1)}):
Requested number of elements in each chunk.
Mutually exclusive with \code{n_chunks} and \code{props}.}

\item{shuffle}{(\code{logical(1)}):
If \code{TRUE}, permutes the order of \code{x} before chunking.}
}
\value{
(integer()`) of chunk indices.
Pass to \code{\link[base:split]{base::split()}} to create a list of chunks.
}
\description{
In case of shuffling and vectors that cannot be chunked evenly,
it is chosen randomly which levels / chunks will receive less elements.
}
\examples{
x = 1:11

ch = chunk(x, n_chunks = 2)
table(ch)
split(x, ch)

ch = chunk(x, n_chunks = 2, shuffle = FALSE)
table(ch)
split(x, ch)

ch = chunk(x, chunk_size = 4)
table(ch)
split(x, ch)
}
