% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\title{Composition of Functions}
\usage{
compose(...)
}
\arguments{
\item{...}{(\code{functions})\cr
Functions to compose.}
}
\value{
(\verb{function()}) which calls the functions provided via \code{...}
in reverse order.
}
\description{
Composes two or more functions into a single function.
The returned function calls all provided functions in reverse order:
The return value of the last function servers as input
for the next to last function, and so on.
}
\examples{
f = compose(function(x) x + 1, function(x) x / 2)
f(10)
}
