% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{config_error}
\alias{config_error}
\title{Configuration Error}
\usage{
config_error(msg, ..., class = NULL)
}
\arguments{
\item{msg}{(\code{character(1)})\cr
Error message.}

\item{...}{(any)\cr
Passed to \code{\link{stopf}}.}

\item{class}{(\code{character})\cr
Additional class(es).}
}
\description{
Configuration error.
Calls \code{\link{stopf}} with class \code{"mlr3ConfigError"}.
}
