% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing.R
\name{count_missing}
\alias{count_missing}
\title{Count Missing Values in a Vector}
\usage{
count_missing(x)
}
\arguments{
\item{x}{\code{\link[=vector]{vector()}}\cr
Supported are logical, integer, double, complex and string vectors.}
}
\value{
(\code{integer(1)}) number of missing values.
}
\description{
Same as \code{sum(is.na(x))}, but without the allocation.
}
\examples{
count_missing(c(1, 2, NA, 4, NA))
}
