% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_join.R
\name{cross_join}
\alias{cross_join}
\title{Cross-Join for data.table}
\usage{
cross_join(dots, sorted = TRUE, unique = FALSE)
}
\arguments{
\item{dots}{(named \code{list()})\cr
Vectors to cross-join.}

\item{sorted}{(\code{logical(1)})\cr
See \code{\link[data.table:J]{data.table::CJ()}}.}

\item{unique}{(\code{logical(1)})\cr
See \code{\link[data.table:J]{data.table::CJ()}}.}
}
\value{
\code{\link[data.table:data.table]{data.table::data.table()}}.
}
\description{
A safe version of \code{\link[data.table:J]{data.table::CJ()}} in case a column is called
\code{sorted} or \code{unique}.
}
\examples{
cross_join(dots = list(sorted = 1:3, b = letters[1:2]))
}
