% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_you_mean.R
\name{did_you_mean}
\alias{did_you_mean}
\title{Selects closest suggestion from a set of strings}
\usage{
did_you_mean(str, candidates)
}
\arguments{
\item{str}{Given string}

\item{candidates}{(\code{character})}
}
\value{
(\code{character(1)}). Either an element from \code{candidates} or the empty string if no close match is found.
}
\description{
Helps to suggest something from a list of names, if the user mistyped in an argcheck.
}
