% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dribble.R
\name{dribble}
\alias{dribble}
\title{Construct a data.table in a row-wise fashion

Similar to \CRANpkg{tibble} function \code{tribble}, this function
allows to construct tabular data in a row-wise fashion.

The first arguments passed as formula will be interpreted as column names.
The remaining arguments will be put into the resulting table.}
\usage{
dribble(...)
}
\arguments{
\item{...}{Arguments: Column names first as formula, then the tabular data.}
}
\value{
\code{\link[`data.table:data.table]{`data.table::data.table()}}.
}
\description{
Construct a data.table in a row-wise fashion

Similar to \CRANpkg{tibble} function \code{tribble}, this function
allows to construct tabular data in a row-wise fashion.

The first arguments passed as formula will be interpreted as column names.
The remaining arguments will be put into the resulting table.
}
\examples{
dribble(
  ~a,  ~b,
  1,   "a",
  2,   "b"
)
}
