% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vars.R
\name{extract_vars}
\alias{extract_vars}
\title{Extract Variables from a Formula}
\usage{
extract_vars(f)
}
\arguments{
\item{f}{:: \code{formula()}.}
}
\value{
(\code{list()}) with elements \code{"lhs"} and \code{"rhs"}, both \code{character()}.
}
\description{
Given a \code{\link[=formula]{formula()}} \code{f}, returns all variables used on the left-hand side and
right-hand side of the formula.
}
\examples{
extract_vars(Species ~ Sepal.Width + Sepal.Length)
extract_vars(Species ~ .)
}
