% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_c.R
\name{fct_c}
\alias{fct_c}
\title{Combine factor vectors}
\usage{
fct_c(...)
}
\arguments{
\item{...}{(two or more objects of type \code{factor()} or \code{character()}).}
}
\value{
(\code{factor()}).
}
\description{
Takes multiple vectors of type \code{factor} or \code{character} and combines them
into a single factor. Levels are merged.
}
\examples{
x = factor(c("a", "b", "a"))
y = factor(c("b", "a", "a"))
z = factor(c("a", "b", "c"))

fct_c(x, y)
fct_c(x, z)
}
