% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seed.R
\name{get_seed}
\alias{get_seed}
\title{Get the random seed}
\usage{
get_seed()
}
\value{
\code{integer()}. Depends on the \code{\link[base:RNGkind]{base::RNGkind()}}.
}
\description{
Retrieves the current random seed (".Random.seed" in the global environment),
and initializes the RNG first, if necessary.
}
\examples{
str(get_seed())
}
