% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr_detect.R
\name{has_element}
\alias{has_element}
\title{Does a list contain an object?}
\usage{
has_element(.x, .y)
}
\arguments{
\item{.x}{(\code{list()} or atomic \code{vector}).}

\item{.y}{(any):
Object to test for}
}
\description{
Simply checks if a list contains a given object.
\itemize{
\item NB1: Objects are compared with identity.
\item NB2: Only use this on lists with complex objects, for simpler structures
there are faster operations.
\item NB3: Clones of R6 objects are not detected.
}
}
