% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashed_sets.R
\name{hashed_sets}
\alias{hashed_sets}
\alias{wunique}
\alias{set_unique_flag}
\alias{set_union}
\alias{set_equal}
\alias{set_intersect}
\alias{set_diff}
\title{Hashing Set Functions}
\usage{
wunique(x, is_unique = FALSE)

set_unique_flag(x)

set_union(x, y)

set_equal(x, y)

set_intersect(x, y)

set_diff(x, y)
}
\arguments{
\item{x}{(\code{vector()}).}

\item{is_unique}{(\code{logical(1)}). Skips calculation of unique values if set to \code{TRUE}.
Used to pass external information about uniqueness down to \code{wunique()}.}

\item{y}{(\code{vector()}).}
}
\description{
The following attributes are attached to atomic input vectors to speed up set operations:
\itemize{
\item \code{.match.hash} is an external pointer to a hash table constructed by \code{\link[fastmatch:fmatch]{fastmatch::fmatch()}}.
\item \code{.unique} is a single logical value which stores that the vector is know to only have unique values.
}

The following functions take advantage of the cached attributes:
\itemize{
\item \code{set_unique_flag()} sets the \code{.unique} attribute by reference.
\item \code{wunique()} is a wrapper around \code{\link[base:unique]{base::unique()}} which either skips computation of unique values
if the \code{.unique} flag is found, or calculates unique values and sets the flag.
\item \code{set_union()}, \code{set_equal()}, \code{set_intersect()} and \code{set_diff()} are the hashed replacement
functions for \code{\link[=union]{union()}}, \code{\link[=setequal]{setequal()}}, \code{\link[=intersect]{intersect()}} and \code{\link[=setdiff]{setdiff()}}.
}
}
