% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Extract ids from a List of Objects}
\usage{
ids(xs)
}
\arguments{
\item{xs}{:: \code{list()}\cr
Every element must have a slot 'id'.}
}
\value{
(\code{character()}).
}
\description{
None.
}
\examples{
xs = list(a = list(id = "foo", a = 1), bar = list(id = "bar", a = 2))
ids(xs)
}
