% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{invoke}
\alias{invoke}
\title{Invoke a function call}
\usage{
invoke(.f, .x = list(), ...)
}
\arguments{
\item{.f}{(\code{\link[=function]{function()}}): Function to call.}

\item{.x}{(\code{\link[=list]{list()}}): List of function arguments passed to \code{.f}.}

\item{...}{: Additional function arguments passed to \code{.f}.
These arguments will be concatenated to the arguments provied via \code{.x}.}
}
\description{
An alternative interface for \code{\link[=do.call]{do.call()}}, similar to the function in \pkg{purrr}.
It is best practice to pass all arguments named.
}
\examples{
invoke(mean, .x = list(x = 1:10))
invoke(mean, .x = list(1:10), na.rm = TRUE)
}
