% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scalar_na.R
\name{is_scalar_na}
\alias{is_scalar_na}
\title{Check if an argument is a single scalar value}
\usage{
is_scalar_na(x)
}
\arguments{
\item{x}{Argument to check.}
}
\value{
(\code{logical(1)}).
}
\description{
Check if an argument is a single scalar value
}
