% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Loads and Returns a Single Data Set}
\usage{
load_dataset(id, package, keep_rownames = FALSE)
}
\arguments{
\item{id}{(\code{character(1)}): Name of the data set.}

\item{package}{(\code{character(1)}): Package to load the data set from.}

\item{keep_rownames}{(\code{logical(1)}): Keep possible row names (default: \code{FALSE}).}
}
\description{
Loads a data set with name \code{id} from package \code{package} and returns it.
}
\examples{
load_dataset("iris", "datasets")
}
