% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info_learner.R
\name{meta_info_learner}
\alias{meta_info_learner}
\title{Creates Markdown Code for Manual Pages Of Learners}
\usage{
meta_info_learner(id)
}
\arguments{
\item{id}{(\code{character(1)})\cr
Id of the learner.}
}
\value{
\code{character()} with markdown code.
}
\description{
Internal function to generate the "Meta Information" section of learners.
}
\keyword{Internal}
