% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info_task.R
\name{meta_info_task}
\alias{meta_info_task}
\title{Creates Markdown Code for Manual Pages Of Tasks}
\usage{
meta_info_task(id)
}
\arguments{
\item{id}{(\code{character(1)})\cr
Id of the tasks.}
}
\value{
\code{character()} with markdown code.
}
\description{
Internal function to generate the "Meta Information" section of tasks.
}
\keyword{Internal}
