% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mlr3Component.R
\name{mlr3component_deprecation_msg}
\alias{mlr3component_deprecation_msg}
\title{Deprecation Message related to the \code{Mlr3Component} Class}
\usage{
mlr3component_deprecation_msg(msg)
}
\arguments{
\item{msg}{(\code{character(1)})
Message to print.}
}
\description{
Will give different messages depending on deprecation progression and will be more agressive in tests than
interactively.
}
\keyword{internal}
