% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3misc-package}
\alias{mlr3misc}
\alias{mlr3misc-package}
\title{mlr3misc: Helper Functions for 'mlr3'}
\description{
Frequently used helper functions and assertions
    used in 'mlr3' and its companion packages. Comes with helper functions
    for functional programming, for printing, to work with 'data.table',
    as well as some generally useful 'R6' classes. This package also
    supersedes the package 'BBmisc'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3misc.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3misc}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3misc/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

}
