% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callback.R
\docType{data}
\name{mlr_callbacks}
\alias{mlr_callbacks}
\title{Dictionary with Callbacks}
\format{
An object of class \code{Dictionary} (inherits from \code{R6}) of length 13.
}
\usage{
mlr_callbacks
}
\description{
This dictionary contains predefined callbacks.
As a convention, the key should start with the name of the package, i.e. package.callback
The elements can be retried using \code{cllb} or \code{cllbs}.
}
\keyword{datasets}
