% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stop_with_class}
\alias{stop_with_class}
\title{Throw Error}
\usage{
stop_with_class(cond)
}
\arguments{
\item{cond}{(\code{error})\cr
The error object.}
}
\description{
Throws an error using \code{\link[cli:cli_abort]{cli::cli_abort}} with the error class included in the message.
}
