% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{mlr_conditions}
\alias{mlr_conditions}
\alias{error_config}
\alias{error_timeout}
\alias{mlr3_error}
\alias{mlr3_warning}
\alias{warning_config}
\title{Error Classes}
\usage{
error_config(msg, ..., class = NULL, silent = FALSE)

error_timeout(silent = FALSE)

mlr3_error(msg, ..., class = NULL)

mlr3_warning(msg, ..., class = NULL)

warning_config(msg, ..., class = NULL, silent = FALSE)
}
\arguments{
\item{msg}{(\code{character(1)})\cr
Error message.}

\item{...}{(any)\cr
Passed to \code{\link[=sprintf]{sprintf()}}.}

\item{class}{(\code{character})\cr
Additional class(es).}

\item{silent}{(\code{logical(1)})\cr
If \code{TRUE}, the condition object is returned.}
}
\description{
Error classes for mlr3.
Currently supported classes:
}
