% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_wrap.R
\name{msg_wrap}
\alias{msg_wrap}
\alias{cat_wrap}
\title{Wrap Strings in Messages}
\usage{
msg_wrap(str, width = 0.9 * getOption("width"))

cat_wrap(str, width = 0.9 * getOption("width"))
}
\arguments{
\item{str}{:: \code{character()}\cr
Vector of strings.}

\item{width}{:: \code{integer(1)}\cr
Width of the output.}
}
\value{
(\code{character()}).
}
\description{
Pretty printing for \code{\link[=message]{message()}} and \code{\link[=cat]{cat()}}.
}
\examples{
msg_wrap("This is a somewhat long message call with a line break
  in it. The output will be wrapped nicely in the console.
  Whitespace will be stripped.")

cat_wrap("This is a somewhat long message call with a line break
  in it. The output will be wrapped nicely in the console.
  Whitespace will be stripped.")
}
\seealso{
\code{\link[=str_indent]{str_indent()}}
}
