% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_vector.R
\name{named_vector}
\alias{named_vector}
\title{Create a Named Vector}
\usage{
named_vector(nn = character(0L), init = NA)
}
\arguments{
\item{nn}{(\code{character()})\cr
Names of new vector}

\item{init}{(\code{atomic})\cr
All vector elements are initialized to this value.}
}
\value{
(named \code{vector()}).
}
\description{
Creates a simple atomic vector with \code{init} as values.
}
\examples{
named_vector(c("a", "b"), NA)
named_vector(character())
}
