% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_help.R
\name{open_help}
\alias{open_help}
\title{Opens a Manual Page}
\usage{
open_help(man)
}
\arguments{
\item{man}{(\code{character(1)})\cr
Manual page to open in "package::topic" syntax.}
}
\value{
Nothing.
}
\description{
Simply opens a manual page specified in "package::topic" syntax.
}
