% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_cbind.R
\name{rcbind}
\alias{rcbind}
\title{Bind Columns by Reference}
\usage{
rcbind(x, y)
}
\arguments{
\item{x}{:: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
\code{data.table()} to add columns to.}

\item{y}{:: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
\code{data.table()} to take columns from.}
}
\value{
Updated \code{x} (\code{data.table}).
}
\description{
Performs \code{\link[base:cbind]{base::cbind()}} on \link[data.table:data.table]{data.tables}, possibly by reference.
}
