% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_info.R
\name{rd_info}
\alias{rd_info}
\alias{rd_format_range}
\alias{rd_format_string}
\alias{rd_format_packages}
\title{Helpers to Create Manual Pages}
\usage{
rd_info(obj, ...)

rd_format_range(lower, upper)

rd_format_string(str, quote = c("\\\\dQuote{", "}"))

rd_format_packages(packages)
}
\arguments{
\item{obj}{(\code{any})\cr
Object of the respective class.}

\item{...}{(\verb{any)})\cr
Additional arguments.}

\item{lower}{(\code{numeric(1)})\cr
Lower bound.}

\item{upper}{(\code{numeric(1)})\cr
Upper bound.}

\item{str}{(\code{character()})\cr
Vector of strings.}

\item{quote}{(\code{character()})\cr
Quotes to use around each element of \code{x}.

Will be replicated to lenght 2.}

\item{packages}{(\code{character()})\cr
Vector of package names.}
}
\value{
\code{character()}, possibly with markdown code.
}
\description{
\code{rd_info()} is an internal generic to generate Rd or markdown code to be used in manual pages.
\code{rd_format_string()} and \code{rd_format_range()} are string functions to assist generating
proper Rd code.
}
\keyword{Internal}
