% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle_vector.R
\name{recycle_vectors}
\alias{recycle_vectors}
\title{Recycle List of Vectors to Common Length}
\usage{
recycle_vectors(.x)
}
\arguments{
\item{.x}{(\code{list()}).}
}
\value{
(\code{list()}) with vectors of same size.
}
\description{
Repeats all vectors of a list \code{.x} to the length of the longest vector
using \code{\link[=rep]{rep()}} with argument \code{length.out}.
This operation will only work
if the length of the longest vectors is an integer multiple of all shorter
vectors, and will throw an exception otherwise.
}
\examples{
recycle_vectors(list(a = 1:3, b = 2))
}
