% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_namespaces.R
\name{require_namespaces}
\alias{require_namespaces}
\title{Require multiple packages}
\usage{
require_namespaces(pkgs,
  msg = "The following packages could not be loaded: \%s")
}
\arguments{
\item{pkgs}{(\code{character()}): Packages to load.}

\item{msg}{(\code{character(1)}): Message to print on error. \code{"\%s"} is placeholder for the list of packages.}
}
\description{
Packages are loaded (not attached) via \code{\link[base:requireNamespace]{base::requireNamespace()}}.
If at least on package can not be loaded, an exception is raised.
}
\examples{
require_namespaces("mlr3misc")
}
