% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{sequence_helpers}
\alias{sequence_helpers}
\alias{seq_row}
\alias{seq_col}
\alias{seq_len0}
\alias{seq_along0}
\title{Sequence Construction Helpers}
\usage{
seq_row(x)

seq_col(x)

seq_len0(n)

seq_along0(x)
}
\arguments{
\item{x}{:: \code{any}\cr
Arbitrary object. Used to query its rows, cols or length.}

\item{n}{:: \code{integer(1)}\cr
Length of the sequence.}
}
\description{
\code{seq_row()} creates a sequence along the number of rows of \code{x},
\code{seq_col()} a sequence along the number of columns of \code{x}.
\code{seq_len0()} and \code{seq_along0()} are the 0-based counterparts to \code{\link[base:seq_len]{base::seq_len()}} and
\code{\link[base:seq_along]{base::seq_along()}}.
}
\examples{
seq_len0(3)
}
