% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\name{set_class}
\alias{set_class}
\title{A wrapper for \code{class(x) = classes}.}
\usage{
set_class(x, classes)
}
\arguments{
\item{x}{(any):
Your object.}

\item{classes}{(\code{character(1)}):
Vector of new classes.}
}
\value{
Changed object \code{x}.
}
\description{
A wrapper for \code{class(x) = classes}.
}
\examples{
set_class(list(), c("foo1", "foo2"))
}
