% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_attr.R
\name{set_names}
\alias{set_names}
\title{Wrapper for setNames}
\usage{
set_names(x, nm = x, ...)
}
\arguments{
\item{x}{(\code{vector}).}

\item{nm}{(\code{character()} | \code{function()}).}

\item{...}{(any). Passed down to \code{nm} if it is a function.}
}
\value{
\code{x} with names set.
}
\description{
Sets the names of \code{x} to \code{nm}.
If \code{nm} is a function, it is used to transform the already existing names of \code{x}.
}
