% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_indent.R
\name{str_indent}
\alias{str_indent}
\title{Indent Strings}
\usage{
str_indent(initial, str, width = 0.9 * getOption("width"), exdent = 2L,
  ...)
}
\arguments{
\item{initial}{(\code{character(1)}):
Initial string, passed to \code{\link[=strwrap]{strwrap()}}.}

\item{str}{(\code{character()}): Vector of strings.}

\item{width}{(\code{integer(1)}):
Width of the output.}

\item{exdent}{(\code{integer(1)}):
Indentation of subsequent lines in paragraph.}

\item{...}{(any): Additional parameters passed to \code{\link[=str_collapse]{str_collapse()}}.}
}
\value{
(\code{character()}).
}
\description{
Indent Strings
}
\examples{
cat(str_indent("Letters:", str_collapse(letters), width = 25), sep = "\\n")
}
