% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_trunc.R
\name{str_trunc}
\alias{str_trunc}
\title{Truncate Strings}
\usage{
str_trunc(str, width = 0.9 * getOption("width"), ellipsis = "[...]")
}
\arguments{
\item{str}{:: \code{character()}\cr
Vector of strings.}

\item{width}{:: \code{integer(1)}\cr
Width of the output.}

\item{ellipsis}{:: \code{character(1)}\cr
If the string has to be shortened, this is signaled by appending \code{ellipsis} to \code{str}. Default is \code{"[...]"}.}
}
\value{
(\code{character()}).
}
\description{
\code{str_trunc()} truncates a string to a given width.
}
\examples{
str_trunc("This is a quite long string", 20)
}
