% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stri_trunc.R
\name{stri_trunc}
\alias{stri_trunc}
\title{Truncate a string to a given length.}
\usage{
stri_trunc(str, width, ellipsis = "[...]")
}
\arguments{
\item{str}{[\code{character()}]:\cr
Vector of strings.}

\item{width}{[\code{integer(1)}]:\cr
Absolute length the string should be truncated to, including \code{ellipsis}.
Note that you cannot truncate to a shorter length than \code{ellipsis}.}

\item{tail}{[\code{character(1)}]:\cr
If the string has to be shortened at least 1 character, the last characters
will be \code{ellipsis}. Default is \code{" [...]"}.}
}
\value{
[\code{character(1)}].
}
\description{
Truncate a string to a given length.
}
\examples{
stri_trunc("This is a quite long string", 20)
}
