% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_collapse.R, R/str_indent.R, R/str_trunc.R
\name{str_collapse}
\alias{str_collapse}
\alias{str_indent}
\alias{string_helpers}
\alias{str_trunc}
\title{Miscellaneous string helper functions}
\usage{
str_collapse(str, sep = ", ", quote = "", n = Inf,
  ellipsis = "[...]")

str_indent(initial, str, width = 0.9 * getOption("width"), ...)

str_trunc(str, width = 0.9 * getOption("width"), ellipsis = "[...]")
}
\arguments{
\item{str}{(\code{character()}): Vector of strings.}

\item{sep}{(\code{character(1)}):\cr
String used to collapse the elements of \code{x}.}

\item{quote}{(\code{character(1)}):\cr
Quotes to use around each element of \code{x}.}

\item{n}{(\code{integer(1)}):\cr
Number of elements to keep from \code{x}. See \code{\link[utils:head]{utils::head()}}.}

\item{ellipsis}{(\code{character(1)}): If the string has to be shortened, this is signaled by appending \code{ellipsis} to \code{str}. Default is \code{" [...]"}.}

\item{initial}{(\code{character(1)}):
Initial string, passed to \code{\link[=strwrap]{strwrap()}}.}

\item{width}{(\code{integer(1)}):
Width of the output.}

\item{...}{(any): Additional parameters passed to \code{\link[=str_collapse]{str_collapse()}} to
collapse \code{str}.}
}
\value{
(\code{character(1)}).
}
\description{
\itemize{
\item \code{str_trunc()} truncates a string to a given width.
\item \code{str_collapse()} returns the first elements of a character vector.
\item \code{str_paragraph()} formats the text in \code{str} with \code{\link[=strwrap]{strwrap()}}.
}
}
\examples{
str_trunc("This is a quite long string", 20)
str_collapse(letters, quote = "'", n = 5)
message(str_indent("Some letters:", str_collapse(letters), width = 25))
}
