% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_srcrefs.R
\name{strip_srcrefs}
\alias{strip_srcrefs}
\title{Strip source references from objects}
\usage{
strip_srcrefs(x, ...)
}
\arguments{
\item{x}{(any)\cr
The object to strip of source references.}

\item{...}{(any)\cr
Additional arguments to the method.}
}
\description{
Source references can make objects unexpectedly large and are undesireable in many situations.
As \CRANpkg{renv} installs packages with the \code{--with-keep.source} option, we sometimes need to remove source references
from objects.
Methods should remove source references from the input, but should otherwise leave the input unchanged.
}
\keyword{internal}
