% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\title{Transpose lists of lists}
\usage{
transpose(.l, ...)
}
\arguments{
\item{.l}{(\code{list()} of \code{list()}).}

\item{\ldots}{Currently unused.}
}
\value{
\code{list()}.
}
\description{
Transposes a list of list, and turns it inside out, similar to the
function \code{transpose()} in package \pkg{purrr}.
}
\examples{
x = list(list(a = 2, b = 3), list(a = 5, b = 10))
str(x)
str(transpose(x))

# list of data frame rows:
transpose(iris[1:2, ])
}
