% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose_list}
\alias{transpose_list}
\title{Transpose lists of lists}
\usage{
transpose_list(.l)
}
\arguments{
\item{.l}{(\code{list()} of \code{list()}).}
}
\value{
\code{list()}.
}
\description{
Transposes a list of list, and turns it inside out, similar to the
function \code{transpose()} in package \CRANpkg{purrr}.
}
\examples{
x = list(list(a = 2, b = 3), list(a = 5, b = 10))
str(x)
str(transpose_list(x))

# list of data frame rows:
transpose_list(iris[1:2, ])
}
