% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{warn_with_class}
\alias{warn_with_class}
\title{Throw Warning}
\usage{
warn_with_class(cond)
}
\arguments{
\item{cond}{(\code{error})\cr
The warning object.}
}
\description{
Throws a warning using \code{\link[cli:cli_abort]{cli::cli_warn}} with the warning class included in the message.
}
