% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_min.R
\name{which_min}
\alias{which_min}
\alias{which_max}
\title{Get the index of the minimum/maximum value, with ties correction

Works similar to \code{\link[base:which.min]{base::which.min()}}/\code{\link[base:which.max]{base::which.max()}}, but corrects for ties.}
\usage{
which_min(x, ties.method = "random")

which_max(x, ties.method = "random")
}
\arguments{
\item{x}{[\code{numeric()}]:\cr
Numeric vector.}

\item{ties.method}{[\code{character(1)}]:\cr
Handling of ties. One of "first", "last" or "random" to return the first index,
the last index, or a random index of the minimum/maximum values.}
}
\value{
[\code{integer(1)}] index of the minimum/maximum value.
Returns [\code{integer(0)}] for empty vectors and vectors with no non-missing values.
}
\description{
Get the index of the minimum/maximum value, with ties correction

Works similar to \code{\link[base:which.min]{base::which.min()}}/\code{\link[base:which.max]{base::which.max()}}, but corrects for ties.
}
\examples{
x = c(2, 3, 1, 3, 5, 1, 1)
which_min(x, ties.method = "first")
which_min(x, ties.method = "last")
which_min(x, ties.method = "random")

which_max(x)
}
