% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callback.R
\name{as_callback}
\alias{as_callback}
\title{Create a Callback}
\usage{
as_callback(id, ...)
}
\arguments{
\item{id}{(\code{character(1)})\cr
Identifier for the new callback.}

\item{...}{(Named list of \verb{function()}s)
Public methods of the \link{Callback}.
The functions must have two arguments named \code{callback} and \code{context}.
The argument names indicate the stage in which the method is called.}
}
\description{
Create a \link{Callback} from a list of functions.
}
