% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callback.R
\name{as_callback}
\alias{as_callback}
\alias{as_callback.Callback}
\alias{as_callbacks}
\alias{as_callbacks.list}
\alias{as_callbacks.Callback}
\title{Convert to a Callback}
\usage{
as_callback(x, ...)

\method{as_callback}{Callback}(x, clone = FALSE, ...)

as_callbacks(x, clone = FALSE, ...)

\method{as_callbacks}{list}(x, clone = FALSE, ...)

\method{as_callbacks}{Callback}(x, clone = FALSE, ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}.}
}
\value{
\link{Callback}.
}
\description{
Convert object to a \link{Callback} or a list of \link{Callback}.
}
