% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_operators.R
\name{check_operators}
\alias{check_operators}
\alias{\%check&&\%}
\alias{\%check||\%}
\title{Logical Check Operators}
\usage{
lhs \%check&&\% rhs

lhs \%check||\% rhs
}
\arguments{
\item{lhs, rhs}{(\verb{check_*}-functions)\cr
\verb{check_*}-functions that return either \code{TRUE} or an error message.}
}
\description{
Logical AND and OR operators for \code{\link{checkmate}} \verb{check_*}-functions.
}
\examples{
check_numeric(x) \%check&&\% check_true(all(x < 0))
check_numeric(x) \%check||\% check_character(x)
}
