% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callback.R
\name{clbk}
\alias{clbk}
\alias{clbks}
\title{Syntactic Sugar for Callback Construction}
\usage{
clbk(.key, ...)

clbks(.keys)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key of the object to construct.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the state of the callback.}

\item{.keys}{(\code{character()})\cr
Keys of the objects to construct.}
}
\description{
Functions to retrieve callbacks from \link{mlr_callbacks} and set parameters in one go.
}
\seealso{
Callback call_back
}
