% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hash.R
\name{hash_input}
\alias{hash_input}
\alias{hash_input.function}
\alias{hash_input.data.table}
\alias{hash_input.list}
\alias{hash_input.default}
\title{Hash Input}
\usage{
hash_input(x)

\method{hash_input}{`function`}(x)

\method{hash_input}{data.table}(x)

\method{hash_input}{list}(x)

\method{hash_input}{default}(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object for which to retrieve the hash input.}
}
\description{
Returns the part of an object to be used to calculate its hash.
}
\section{Methods (by class)}{
\itemize{
\item \code{hash_input(`function`)}: The formals and the body are returned in a \code{list()}.
This ensures that neither the bytecode nor the parent environment are included in the hash.
Source references are removed from the body, so that the hash does not depend on the formatting
of the definition or on its position in a source file.
Primitives have neither formals nor a body and are returned as is.

\item \code{hash_input(data.table)}: The data.table is converted to a regular list and \code{hash_input()} is applied to all elements.
The conversion to a list ensures that keys and indices are not included in the hash.

\item \code{hash_input(list)}: \code{hash_input()} is applied to all elements of the list.

\item \code{hash_input(default)}: Returns the object as is.

}}
