% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hash.R
\name{hash_input}
\alias{hash_input}
\title{Hash Input}
\usage{
hash_input(x)
}
\arguments{
\item{x}{(any)\cr
Object for which to calculate the hash.}
}
\description{
Returns the information of an object to be used to calculate its hash.
}
