% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Short helper to extract ids from a typed R6 list of ID-able objects.}
\usage{
ids(xs)
}
\arguments{
\item{xs}{:: \link{list}\cr
Every element must be R6 and have a slot 'id'.}
}
\description{
None.
}
