% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scalar_na.R
\name{is_scalar_na}
\alias{is_scalar_na}
\title{Check for a Single Scalar Value}
\usage{
is_scalar_na(x)
}
\arguments{
\item{x}{(\code{any})\cr
Argument to check.}
}
\value{
(\code{logical(1)}).
}
\description{
Check for a Single Scalar Value
}
