% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{meta_info_task}
\alias{meta_info_task}
\alias{meta_info_learner}
\alias{meta_info_measure}
\title{Creates Markdown Code for Manual Pages}
\usage{
meta_info_task(obj)

meta_info_learner(obj)

meta_info_measure(obj)
}
\arguments{
\item{obj}{(\code{any})\cr
Object of the respective class.}
}
\value{
\code{character()} with markdown code.
}
\description{
Internal function to generate the "Meta Information" section of for
\CRANpkg{mlr3} tasks, learners or measures.
}
\keyword{Internal}
