% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callback.R
\docType{data}
\name{mlr_callbacks}
\alias{mlr_callbacks}
\title{Dictionary of Callbacks}
\format{
An object of class \code{DictionaryCallbacks} (inherits from \code{Dictionary}, \code{R6}) of length 12.
}
\usage{
mlr_callbacks
}
\description{
A simple \link{Dictionary} storing objects of class \link{Callback}.
Each callback has an associated help page, see \code{mlr_callbacks_[id]}.

This dictionary can get populated with additional callbacks by add-on packages.
As a convention, the key should start with the name of the package, i.e. \code{package.callback}.

For a more convenient way to retrieve and construct learners, see \code{\link[=clbk]{clbk()}}/\code{\link[=clbks]{clbks()}}.
}
\keyword{datasets}
