% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{mlr_conditions}
\alias{mlr_conditions}
\alias{error_config}
\alias{condition_config}
\alias{error_timeout}
\alias{condition_timeout}
\alias{error_mlr3}
\alias{condition_mlr3}
\title{Error Classes}
\usage{
error_config(msg, ..., class = NULL)

condition_config(msg, ..., class = NULL)

error_timeout()

condition_timeout()

error_mlr3(msg, ..., class = NULL)

condition_mlr3(msg, ..., class = NULL)
}
\arguments{
\item{msg}{(\code{character(1)})\cr
Error message.}

\item{...}{(any)\cr
Passed to \code{\link[=sprintf]{sprintf()}}.}

\item{class}{(\code{character})\cr
Additional class(es).}
}
\description{
Error classes for mlr3.
Currently supported classes:
}
