% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_list.R
\name{named_list}
\alias{named_list}
\title{Create a named list, possibly initialized with a certain element.}
\usage{
named_list(nn, init = NULL)
}
\arguments{
\item{nn}{(\code{character()}):\cr
Names of elements.}

\item{init}{(valid R expression):\cr
All list elements are initialized to this value.}
}
\value{
(\code{list()}].
}
\description{
Create a named list, possibly initialized with a certain element.
}
\examples{
named_list(c("a", "b"))
named_list(c("a", "b"), init = 1)
}
