% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_list.R
\name{named_list}
\alias{named_list}
\title{Create a Named List}
\usage{
named_list(nn = character(0L), init = NULL)
}
\arguments{
\item{nn}{:: \code{character()}\cr
Names of new list.}

\item{init}{:: \code{any}\cr
All list elements are initialized to this value.}
}
\value{
(named \code{list()}).
}
\description{
Create a Named List
}
\examples{
named_list(c("a", "b"))
named_list(c("a", "b"), init = 1)
}
