% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printf.R
\name{printf}
\alias{printf}
\alias{catf}
\alias{messagef}
\alias{warningf}
\alias{stopf}
\title{Functions for Formatted Output and Conditions}
\usage{
catf(msg, ..., file = "")

messagef(msg, ...)

warningf(msg, ...)

stopf(msg, ...)
}
\arguments{
\item{msg}{:: \code{character(1)}\cr
Format string passed to \code{\link[base:sprintf]{base::sprintf()}}.}

\item{...}{:: \code{any}\cr
Arguments passed down to \code{\link[base:sprintf]{base::sprintf()}}.}

\item{file}{:: \code{character(1)}\cr
Passed to \code{\link[base:cat]{base::cat()}}.}
}
\description{
\code{catf()}, \code{messagef()}, \code{warningf()} and \code{stopf()} are wrappers around \code{\link[base:cat]{base::cat()}},
\code{\link[base:message]{base::message()}}, \code{\link[base:warning]{base::warning()}} and \code{\link[base:stop]{base::stop()}}, respectively.
}
