% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbind.R
\name{rcbind}
\alias{rcbind}
\title{Bind Columns by Reference}
\usage{
rcbind(x, y)
}
\arguments{
\item{x}{:: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
\code{\link[data.table:data.table]{data.table::data.table()}} to add columns to.}

\item{y}{:: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
\code{\link[data.table:data.table]{data.table::data.table()}} to take columns from.}
}
\value{
(\code{\link[data.table:data.table]{data.table::data.table()}}): Updated \code{x} .
}
\description{
Performs \code{\link[base:cbind]{base::cbind()}} on \link[data.table:data.table]{data.tables}, possibly by reference.
}
