% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\name{set_class}
\alias{set_class}
\title{Set the Class}
\usage{
set_class(x, classes)
}
\arguments{
\item{x}{:: \code{any.}}

\item{classes}{:: \code{character(1)}\cr
Vector of new class names.}
}
\value{
Object \code{x}, with updated class attribute.
}
\description{
Simple wrapper for \code{class(x) = classes}.
}
\examples{
set_class(list(), c("foo1", "foo2"))
}
