% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_collapse.R, R/str_trunc.R
\name{str_collapse}
\alias{str_collapse}
\alias{string_helpers}
\alias{str_trunc}
\title{Miscellaneous string helper functions}
\usage{
str_collapse(str, sep = ", ", quote = "", n = Inf,
  ellipsis = "[...]")

str_trunc(str, width, ellipsis = "[...]")
}
\arguments{
\item{str}{(\code{character()}): Vector of strings.}

\item{sep}{(\code{character(1)}):\cr
String used to collapse the elements of \code{x}.}

\item{quote}{(\code{character(1)}):\cr
Quotes to use around each element of \code{x}.}

\item{n}{(\code{integer(1)}):\cr
Number of elements to keep from \code{x}. See \code{\link[utils:head]{utils::head()}}.}

\item{ellipsis}{(\code{character(1)}): If the string has to be shortened, this is signaled by appending \code{ellipsis} to \code{str}. Default is \code{" [...]"}.}

\item{width}{(\code{integer(1)}):
Absolute length the string should be truncated to, including \code{ellipsis}.
Note that you cannot truncate to a shorter length than \code{ellipsis}.}
}
\value{
(\code{character(1)}).
}
\description{
\itemize{
\item \code{str_trunc()} truncates a string to a given width.
\item \code{str_collapse()} returns the first elements of a character vector.
}
}
\examples{
str_trunc("This is a quite long string", 20)
}
