% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest}
\alias{unnest}
\title{Unnest list data table columns

Transforms each element of a list columns into its own column
by reference.}
\usage{
unnest(x, cols)
}
\arguments{
\item{x}{[\code{data.table}]:\cr
\code{data.table} with columns to unnest.}

\item{cols}{[\code{character()}]:\cr
Column names of list columns to unnest.}
}
\value{
Updated \code{x} [\code{data.table}].
}
\description{
Unnest list data table columns

Transforms each element of a list columns into its own column
by reference.
}
